refactor: implement requested changes
